+++ src/Expression/Value/Builder/DateTimeValueBuilder.php
@@ -11,6 +11,8 @@ use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 use Yiisoft\Db\Schema\Column\ColumnFactoryInterface;
 
+use function Yiisoft\Db\dbTypecast;
+
 /**
  * Builder for {@see DateTimeValue} expressions.
  *
@@ -33,9 +35,8 @@ final class DateTimeValueBuilder implements ExpressionBuilderInterface
 
     public function build(ExpressionInterface $expression, array &$params = []): string
     {
-        $value = $this->columnFactory
-            ->fromType($this->prepareType($expression), $this->prepareInfo($expression))
-            ->dbTypecast($expression->value);
+        $column = $this->columnFactory->fromType($this->prepareType($expression), $this->prepareInfo($expression));
+        $value = dbTypecast($column, $expression->value);
         return $this->queryBuilder->buildValue($value, $params);
     }