+++ src/QueryBuilder/AbstractColumnDefinitionBuilder.php
@@ -12,6 +12,7 @@ use Yiisoft\Db\Schema\Column\EnumColumn;
 
 use function in_array;
 use function strtolower;
+use function Yiisoft\Db\dbTypecast;
 
 /**
  * Builds column definition from {@see ColumnInterface} object. Column definition is a string that represents
@@ -193,7 +194,7 @@ abstract class AbstractColumnDefinitionBuilder implements ColumnDefinitionBuilde
             return '';
         }
 
-        $defaultValue = $column->dbTypecast($column->getDefaultValue());
+        $defaultValue = dbTypecast($column, $column->getDefaultValue());
         $defaultValue = $this->queryBuilder->prepareValue($defaultValue);
 
         if ($defaultValue === '') {