+++ src/Schema/Column/BitColumn.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Schema\Column;
 
 use Yiisoft\Db\Constant\ColumnType;
-use Yiisoft\Db\Expression\ExpressionInterface;
 
 use function is_int;
 
@@ -16,7 +15,7 @@ class BitColumn extends AbstractColumn
 {
     protected const DEFAULT_TYPE = ColumnType::BIT;
 
-    public function dbTypecast(mixed $value): int|string|ExpressionInterface|null
+    public function dbTypecast(mixed $value): int|string|null
     {
         if (is_int($value)) {
             return $value;
@@ -24,7 +23,7 @@ class BitColumn extends AbstractColumn
 
         return match ($value) {
             null, '' => null,
-            default => $value instanceof ExpressionInterface ? $value : (int) $value,
+            default => (int) $value,
         };
     }