+++ src/Schema/Column/BooleanColumn.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Schema\Column;
 
 use Yiisoft\Db\Constant\ColumnType;
-use Yiisoft\Db\Expression\ExpressionInterface;
 
 /**
  * Represents the metadata for a boolean column.
@@ -14,13 +13,13 @@ class BooleanColumn extends AbstractColumn
 {
     protected const DEFAULT_TYPE = ColumnType::BOOLEAN;
 
-    public function dbTypecast(mixed $value): bool|ExpressionInterface|null
+    public function dbTypecast(mixed $value): ?bool
     {
         return match ($value) {
             true => true,
             false => false,
             null, '' => null,
-            default => $value instanceof ExpressionInterface ? $value : (bool) $value,
+            default => (bool) $value,
         };
     }