+++ src/Schema/Column/IntegerColumn.php
@@ -9,7 +9,6 @@ use DateTimeInterface;
 use Stringable;
 use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\Constant\GettypeResult;
-use Yiisoft\Db\Expression\ExpressionInterface;
 
 use function gettype;
 
@@ -20,9 +19,9 @@ class IntegerColumn extends AbstractColumn
 {
     protected const DEFAULT_TYPE = ColumnType::INTEGER;
 
-    public function dbTypecast(mixed $value): int|ExpressionInterface|null
+    public function dbTypecast(mixed $value): ?int
     {
-        /** @var ExpressionInterface|int|null */
+        /** @var int|null */
         return match (gettype($value)) {
             GettypeResult::INTEGER => $value,
             GettypeResult::NULL => null,
@@ -30,7 +29,6 @@ class IntegerColumn extends AbstractColumn
             GettypeResult::DOUBLE => (int) $value,
             GettypeResult::BOOLEAN => $value ? 1 : 0,
             GettypeResult::OBJECT => match (true) {
-                $value instanceof ExpressionInterface => $value,
                 $value instanceof BackedEnum => $value->value === '' ? null : (int) $value->value,
                 $value instanceof DateTimeInterface => $value->getTimestamp(),
                 $value instanceof Stringable => ($val = (string) $value) === '' ? null : (int) $val,