+++ src/Schema/Column/StringColumn.php
@@ -8,7 +8,6 @@ use BackedEnum;
 use Stringable;
 use UnitEnum;
 use Yiisoft\Db\Constant\ColumnType;
-use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Constant\GettypeResult;
 
 use function gettype;
@@ -41,7 +40,6 @@ class StringColumn extends AbstractColumn implements CollatableColumnInterface
             GettypeResult::DOUBLE => (string) $value,
             GettypeResult::BOOLEAN => $value ? '1' : '0',
             GettypeResult::OBJECT => match (true) {
-                $value instanceof ExpressionInterface => $value,
                 $value instanceof BackedEnum => (string) $value->value,
                 $value instanceof UnitEnum => $value->name,
                 $value instanceof Stringable => (string) $value,